+++ tests/testthat/test-ggplot-rect.R
@@ -1,4 +1,3 @@
-context("geom_rect")
 
 expect_traces <- function(gg, n.traces, name) {
   stopifnot(is.numeric(n.traces))