+++ tests/testthat/test-ggplot-resize.R
@@ -1,4 +1,3 @@
-context("ggplotly-resize")
 
 test_that("ggplotly shouldn't populate layout.[width/height] by default", {
   lay <- ggplotly(qplot(1:10))$x$layout