+++ tests/testthat/test-ggplot-smooth.R
@@ -1,4 +1,4 @@
-context("smooth")
+
 
 expect_traces <- function(gg, n.traces, name){
   stopifnot(is.numeric(n.traces))