+++ tests/testthat/test-plotly-color.R
@@ -1,4 +1,3 @@
-context("plotly-color")
 
 expect_traces <- function(p, n.traces, name){
   stopifnot(is.numeric(n.traces))