+++ tests/testthat/test-plotly-group.R
@@ -1,4 +1,4 @@
-context("plotly-group")
+
 
 expect_traces <- function(p, n.traces, name){
   stopifnot(is.numeric(n.traces))