+++ tests/testthat/test-plotly-waterfall.R
@@ -1,4 +1,3 @@
-context("waterfall")
 
 test_that("Simple waterfall works", {
   p <- plot_ly() %>%